+++ packages/tsrx/src/index.js
@@ -133,6 +133,14 @@ export { escape } from './utils/escaping.js';
 // Transform
 export { render_stylesheets as renderStylesheets } from './transform/stylesheet.js';
 export {
+	prepare_stylesheet_for_render as prepareStylesheetForRender,
+	is_style_element as isStyleElement,
+	is_composite_element as isCompositeElement,
+	annotate_with_hash as annotateWithHash,
+	annotate_component_with_hash as annotateComponentWithHash,
+	add_hash_class as addHashClass,
+} from './transform/scoping.js';
+export {
 	convert_source_map_to_mappings as convertSourceMapToMappings,
 	create_volar_mappings_result as createVolarMappingsResult,
 } from './transform/segments.js';
